+++ tests/testthat/testQuantiles.R
@@ -1,10 +1,12 @@
 context("test_phe_quantiles")
 
 # test grouped df field
-df1 <- test_quantiles_g %>% filter(GroupSet == "IndSexReg")
+df1 <- test_quantiles_g %>% filter(GroupSet == "IndSexReg") %>%
+  group_by(IndSexRef, ParentCode)
 
 # test grouped df logical
-df2 <- test_quantiles_g %>% filter(IndSexRef == "90366Female"& GroupSet == "IndSexReg")
+df2 <- test_quantiles_g %>% filter(IndSexRef == "90366Female"& GroupSet == "IndSexReg") %>%
+  group_by(IndSexRef, ParentCode)
 
 # test ungrouped df field
 df3 <- test_quantiles_g %>% filter(GroupSet == "IndSex")
@@ -12,16 +14,18 @@ df3 <- test_quantiles_g %>% filter(GroupSet == "IndSex")
 # test ungrouped df logical
 df4 <- test_quantiles_ug %>% filter(GroupSet == "None")
 
+# test grouped
+df5 <- df4 %>% group_by(GroupSet)
 
 
 #test calculations
 test_that("quantiles calculate correctly",{
-  expect_equal(data.frame(phe_quantile(group_by(df1, ParentCode, add=TRUE),Value,
+  expect_equal(data.frame(phe_quantile(df1,Value,
                             invert = Polarity, inverttype = "field")[15]),
                rename(df1,quantile = QuantileInGrp)[14],
                check.attributes=FALSE, check.names=FALSE,info="test grouped df field")
 
-  expect_equal(data.frame(phe_quantile(group_by(df2, ParentCode, add=TRUE),Value,
+  expect_equal(data.frame(phe_quantile(df2,Value,
                             invert = FALSE))[15:18],
                data.frame(tibble(quantile = df2$QuantileInGrp,
                       nquantiles = 10L,
@@ -29,14 +33,14 @@ test_that("quantiles calculate correctly",{
                       qinverted = "lowest quantile represents lowest values")),
                check.attributes=FALSE, check.names=FALSE,info="test grouped df logical")
 
-  expect_equal(phe_quantile(group_by(df3, IndSexRef, add=TRUE), Value,
+  expect_equal(phe_quantile(df3, Value,
                             invert = Polarity, inverttype = "field", nquantiles = 7L)[15],
                rename(df3,quantile = QuantileInGrp)[14],check.attributes=FALSE,
                check.names=FALSE,info="test ungrouped df field")
 
-  expect_equal(phe_quantile(group_by(df4, GroupSet, add=TRUE), Value, nquantiles = 4L)[15],
+  expect_equal(phe_quantile(df5, Value, nquantiles = 4L)[15],
                rename(df4,quantile = QuantileInGrp)[14],
-               check.attributes=FALSE, check.names=FALSE,info="test ungrouped df logical")
+               check.attributes=FALSE, check.names=FALSE,info="test nquantiles")
 
   expect_equal(phe_quantile(df4, Value, nquantiles = 4L)[15],
                rename(df4,quantile = QuantileInGrp)[14],