+++ R/PHEindicatormethods.R
@@ -1,6 +1,9 @@
 #' PHEindicatormethods: A package for performing standard statistics for public
 #' health indicators
 #'
+#' A package for performing standard statistics for public
+#' health indicators.
+#'
 #' @docType package
 #' @name PHEindicatormethods
 NULL
@@ -10,6 +13,7 @@ globalVariables(c("a_vals",
                   "b_sqrt_a",
                   "b_vals",
                   "CI_params",
+                  "dths_used",
                   "ei",
                   "esp2013",
                   "estimate",
@@ -25,10 +29,12 @@ globalVariables(c("a_vals",
                   "method",
                   "model",
                   "naflag",
+                  "na_flag",
                   "nrefpop_calc",
                   "observed",
                   "p_2b_removed",
                   "p.value",
+                  "pops_used",
                   "qchisq",
                   "qi_2b_removed",
                   "qinverted",